+++ src/video_output/vout_subpictures.c
@@ -905,14 +905,14 @@ static void SpuRenderRegion(spu_t *spu,
 
         /* We suppose DVD palette here */
         new_palette.i_entries = 4;
-        for (int i = 0; i < 4; i++)
+        for (size_t i = 0; i < 4; i++)
         {
             new_palette.palette[i] = sys->palette.palette[i];
             b_opaque |= (new_palette.palette[i].rgba.a > 0x00);
         }
 
         if (old_palette->i_entries == new_palette.i_entries) {
-            for (int i = 0; i < old_palette->i_entries; i++)
+            for (size_t i = 0; i < old_palette->i_entries; i++)
             {
                 changed_palette |= memcmp(&old_palette->palette[i], &new_palette.palette[i], 4);
                 b_old_opaque |= (old_palette->palette[i].rgba.a > 0x00);
@@ -1309,7 +1309,7 @@ static void UpdateSPU(spu_t *spu, const vlc_spu_highlight_t *hl)
     if (hl->palette.i_entries == 4) /* XXX: Only DVD palette for now */
         memcpy(&sys->palette, &hl->palette, sizeof(sys->palette));
 
-    msg_Dbg(spu, "crop: %i,%i,%i,%i, palette forced: %i",
+    msg_Dbg(spu, "crop: %i,%i,%i,%i, palette forced: %zu",
             sys->crop.x, sys->crop.y,
             sys->crop.width, sys->crop.height,
             sys->palette.i_entries);
